moved testcases for sqlite to its own file
