+++ connection_details_test.go
@@ -84,115 +84,3 @@ func Test_ConnectionDetails_Finalize_NoDB_NoURL(t *testing.T) {
 	err := cd.Finalize()
 	r.Error(err)
 }
-
-func Test_ConnectionDetails_Finalize_SQLite_URL_Only(t *testing.T) {
-	r := require.New(t)
-
-	cd := &ConnectionDetails{
-		URL: "sqlite3:///tmp/foo.db",
-	}
-	err := cd.Finalize()
-	r.NoError(err)
-	r.Equal("sqlite3", cd.Dialect, "given dialect: N/A")
-	r.Equal("/tmp/foo.db", cd.Database, "given url: sqlite3:///tmp/foo.db")
-	r.Equal("", cd.Host)
-	r.Equal("", cd.Password)
-	r.Equal("", cd.Port)
-	r.Equal("", cd.User)
-}
-
-func Test_ConnectionDetails_Finalize_SQLite_SynURL_Only(t *testing.T) {
-	r := require.New(t)
-
-	cd := &ConnectionDetails{
-		URL: "sqlite:///tmp/foo.db",
-	}
-	err := cd.Finalize()
-	r.NoError(err)
-	r.Equal("sqlite3", cd.Dialect, "given dialect: N/A")
-	r.Equal("/tmp/foo.db", cd.Database, "given url: sqlite:///tmp/foo.db")
-	r.Equal("", cd.Host)
-	r.Equal("", cd.Password)
-	r.Equal("", cd.Port)
-	r.Equal("", cd.User)
-}
-
-func Test_ConnectionDetails_Finalize_SQLite_Dialect_URL(t *testing.T) {
-	r := require.New(t)
-	cd := &ConnectionDetails{
-		Dialect: "sqlite3",
-		URL:     "sqlite3:///tmp/foo.db",
-	}
-	err := cd.Finalize()
-	r.NoError(err)
-	r.Equal("sqlite3", cd.Dialect, "given dialect: sqlite3")
-	r.Equal("/tmp/foo.db", cd.Database, "given url: sqlite3:///tmp/foo.db")
-	r.Equal("", cd.Host)
-	r.Equal("", cd.Password)
-	r.Equal("", cd.Port)
-	r.Equal("", cd.User)
-}
-
-func Test_ConnectionDetails_Finalize_SQLite_Dialect_SynURL(t *testing.T) {
-	r := require.New(t)
-	cd := &ConnectionDetails{
-		Dialect: "sqlite3",
-		URL:     "sqlite:///tmp/foo.db",
-	}
-	err := cd.Finalize()
-	r.NoError(err)
-	r.Equal("sqlite3", cd.Dialect, "given dialect: sqlite3")
-	r.Equal("/tmp/foo.db", cd.Database, "given url: sqlite:///tmp/foo.db")
-	r.Equal("", cd.Host)
-	r.Equal("", cd.Password)
-	r.Equal("", cd.Port)
-	r.Equal("", cd.User)
-}
-
-func Test_ConnectionDetails_Finalize_SQLite_Synonym_URL(t *testing.T) {
-	r := require.New(t)
-	cd := &ConnectionDetails{
-		Dialect: "sqlite",
-		URL:     "sqlite3:///tmp/foo.db",
-	}
-	err := cd.Finalize()
-	r.NoError(err)
-	r.Equal("sqlite3", cd.Dialect, "given dialect: sqlite")
-	r.Equal("/tmp/foo.db", cd.Database, "given url: sqlite3:///tmp/foo.db")
-	r.Equal("", cd.Host)
-	r.Equal("", cd.Password)
-	r.Equal("", cd.Port)
-	r.Equal("", cd.User)
-}
-
-func Test_ConnectionDetails_Finalize_SQLite_Synonym_SynURL(t *testing.T) {
-	r := require.New(t)
-	cd := &ConnectionDetails{
-		Dialect: "sqlite",
-		URL:     "sqlite:///tmp/foo.db",
-	}
-	err := cd.Finalize()
-	r.NoError(err)
-	r.Equal("sqlite3", cd.Dialect, "given dialect: sqlite")
-	r.Equal("/tmp/foo.db", cd.Database, "given url: sqlite:///tmp/foo.db")
-	r.Equal("", cd.Host)
-	r.Equal("", cd.Password)
-	r.Equal("", cd.Port)
-	r.Equal("", cd.User)
-}
-
-func Test_ConnectionDetails_Finalize_SQLite_Synonym_Path(t *testing.T) {
-	r := require.New(t)
-	cd := &ConnectionDetails{
-		Dialect:  "sqlite",
-		Database: "./foo.db",
-	}
-	err := cd.Finalize()
-	r.NoError(err)
-	r.Equal("sqlite3", cd.Dialect, "given dialect: sqlite")
-	r.Equal("./foo.db", cd.Database, "given database: ./foo.db")
-	r.Equal("", cd.Host)
-	r.Equal("", cd.Password)
-	r.Equal("", cd.Port)
-	r.Equal("", cd.User)
-}
